+++ test/outline/test_outline_serializer.py
@@ -4,16 +4,20 @@ from fpdf.syntax import DestinationXYZ, PDFString
 
 def test_serialize_outline():
     sections = (
-        OutlineSection("Title 1", level=0, page_number=1, dest=DestinationXYZ(page=1)),
         OutlineSection(
-            "Subtitle 1.1", level=1, page_number=1, dest=DestinationXYZ(page=2)
-        ),
-        OutlineSection("Title 2", level=0, page_number=1, dest=DestinationXYZ(page=3)),
-        OutlineSection(
-            "Subtitle 2.1", level=1, page_number=1, dest=DestinationXYZ(page=4)
+            "Title 1", level=0, page_number=1, dest=DestinationXYZ(page=1, top=0)
         ),
         OutlineSection(
-            "Subtitle 2.2", level=1, page_number=1, dest=DestinationXYZ(page=5)
+            "Subtitle 1.1", level=1, page_number=1, dest=DestinationXYZ(page=2, top=0)
+        ),
+        OutlineSection(
+            "Title 2", level=0, page_number=1, dest=DestinationXYZ(page=3, top=0)
+        ),
+        OutlineSection(
+            "Subtitle 2.1", level=1, page_number=1, dest=DestinationXYZ(page=4, top=0)
+        ),
+        OutlineSection(
+            "Subtitle 2.2", level=1, page_number=1, dest=DestinationXYZ(page=5, top=0)
         ),
     )
     assert (
@@ -80,11 +84,19 @@ endobj"""
 
 def test_serialize_outline_with_headless_hierarchy():  # issues 239
     sections = (
-        OutlineSection("?-1", level=1, page_number=2, dest=DestinationXYZ(page=2)),
-        OutlineSection("?-1-1", level=2, page_number=2, dest=DestinationXYZ(page=2)),
-        OutlineSection("1", level=0, page_number=2, dest=DestinationXYZ(page=2)),
-        OutlineSection("1-1", level=1, page_number=2, dest=DestinationXYZ(page=2)),
-        OutlineSection("1-1-1", level=2, page_number=2, dest=DestinationXYZ(page=2)),
+        OutlineSection(
+            "?-1", level=1, page_number=2, dest=DestinationXYZ(page=2, top=0)
+        ),
+        OutlineSection(
+            "?-1-1", level=2, page_number=2, dest=DestinationXYZ(page=2, top=0)
+        ),
+        OutlineSection("1", level=0, page_number=2, dest=DestinationXYZ(page=2, top=0)),
+        OutlineSection(
+            "1-1", level=1, page_number=2, dest=DestinationXYZ(page=2, top=0)
+        ),
+        OutlineSection(
+            "1-1-1", level=2, page_number=2, dest=DestinationXYZ(page=2, top=0)
+        ),
     )
     assert (
         serialize_outline(sections, first_object_id=6)
@@ -151,7 +163,7 @@ def test_serialize_outline_without_hex_encoding():  # issue-458
     try:
         sections = (
             OutlineSection(
-                "Title", level=0, page_number=1, dest=DestinationXYZ(page=1)
+                "Title", level=0, page_number=1, dest=DestinationXYZ(page=1, top=0)
             ),
         )
         assert (