+++ test/test_annotations.py
@@ -76,7 +76,7 @@ def test_goto_action(tmp_path):
     x, y, text = 80, 140, "GoTo action"
     pdf.text(x=x, y=y, txt=text)
     pdf.add_action(
-        GoToAction(dest=DestinationXYZ(page=2).as_str(pdf)),
+        GoToAction(dest=DestinationXYZ(page=2, top=pdf.h_pt)),
         x=x,
         y=y - pdf.font_size,
         w=pdf.get_string_width(text),
@@ -96,7 +96,7 @@ def test_goto_next_page_chained(tmp_path):
     pdf.cell(txt="Page 1 (first page)")
     pdf.add_action(
         GoToAction(
-            DestinationXYZ(page=1).as_str(pdf),
+            DestinationXYZ(page=1, top=pdf.h_pt),
             next_action=pdf_ref(object_id_for_page(pdf.page + 1)),
         ),
         x=0,
@@ -108,7 +108,7 @@ def test_goto_next_page_chained(tmp_path):
     pdf.cell(txt="Page 2")
     pdf.add_action(
         GoToAction(
-            DestinationXYZ(page=1).as_str(pdf),
+            DestinationXYZ(page=1, top=pdf.h_pt),
             next_action=pdf_ref(object_id_for_page(pdf.page + 1)),
         ),
         x=0,
@@ -131,7 +131,7 @@ def test_infinite_loop_with_goto_action(tmp_path):
     pdf.add_page()
     pdf.add_action(
         GoToAction(
-            DestinationXYZ(page=1).as_str(pdf),
+            DestinationXYZ(page=1, top=pdf.h_pt),
             next_action=pdf_ref(object_id_for_page(pdf.page)),
         ),
         x=0,
@@ -148,7 +148,7 @@ def test_goto_remote_action(tmp_path):
     pdf.add_page()
     x, y, text = 80, 140, "GoTo-Remote action"
     pdf.text(x=x, y=y, txt=text)
-    dest = DestinationXYZ(page=1, page_as_obj_id=False).as_str(pdf)
+    dest = DestinationXYZ(page=1, top=pdf.h_pt, page_as_obj_id=False)
     pdf.add_action(
         GoToRemoteAction("goto_action.pdf", dest=dest),
         x=x,